Completed Game
I completed the game by incorporating another page connected to the quiz window, which indicates the score and a comment on the score (made with an if statement) on how well the user did with the questions. I also completed the learning pages by incorporating images of the theory work, and designed the buttons by making them more colourful and larger in size.